refactor proof pack report input router
